Update CoinBalanceResponse coinObjectCount type

+++ models/read_coin.go
@@ -19,7 +19,7 @@ type CoinLockedBalance struct {
 
 type CoinBalanceResponse struct {
 	CoinType        string            `json:"coinType"`
-	CoinObjectCount int               `json:"coinObjectCount"`
+	CoinObjectCount uint64            `json:"coinObjectCount"`
 	TotalBalance    string            `json:"totalBalance"`
 	LockedBalance   CoinLockedBalance `json:"lockedBalance"`
 }
